VRDancing-Discord-Bot-DB/requirements.txt
discord.py
pillow
asyncpg